Fix AWS deployment on Swift 5.4
